requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
cffi==1.15.1
charset-normalizer==2.0.12
cryptography==37.0.4
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.2.4
Django==4.0.4
django-allauth==0.51.0
django-cors-headers==3.12.0
django-crontab==0.7.1
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
idna==3.3
oauthlib==3.2.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
urllib3==1.26.9
whitenoise==6.2.0
